# swagger.yaml
openapi: 3.0.0
info:
  title: Mi E-Commerce
  version: 2.0.0
  description: |-
    Esta API esta orientada a interactuar con el front de una pagina web de E-Commerce del grupo Cencosud SA.
    Git Repo: https://github.com/rodrigobenta/segundoSprint  
servers:
  - url: http://localhost:3000/api/v2
    description: Entorno Desarrollo
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
paths:
...
  /products:
    get:
      tags: 
        - productos
      summary: Lista todos los productos 
      security:
        - bearerAuth: []
      parameters:
        - name: category
          in: query
          schema:
            type: string
          required: false
          description: busca los productos segun su categoria, sino, por defecto, lista todos los productos
      
      responses: 
        '200':
          description: Listado de productos
          content:
            application/json:
              schema:
                properties:
                  list:
                    type: array
                    items:
                      type: object
                      properties:
                        id:
                          type: number
                        title:
                          type: string
                        price:
                          type: number
                        description:
                          type: string
                        image:
                          type: string
                        gallery:
                          type: array
                          items:
                            type: object 
                            properties:
                              picture_id:
                                type: number
                              picture_url:
                                type: string
                              description:
                                type: string
                        category:
                          type: string        
                        mostwanted:
                          type: boolean
                        stock:
                          type: number

        '404':
          description: Productos con dicha categoria no encontrados
          content:
            application/json:
              schema:
                properties:
                  msg:
                    type: string                      
      
        
        '500':
          description: Server error
          content:
            application/json:
              schema:
                properties:
                  Mensaje: 
                    type: string
    post:
      tags: 
        - productos
      summary: Crea un producto 
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              properties:
                title:
                  type: string
                stock:
                  type: number
                description:
                  type: string
                fk_id_category:
                  type: string
                mostwanted:
                  type: number

      
      responses: 
        '201':
          description: Producto creado con exito
          content:
            application/json:
              schema:
                properties:
                  list:
                    type: object
                    properties:
                      id:
                        type: number
                      title:
                        type: string
                      price:
                        type: number
                      description:
                        type: string
                      image:
                        type: string
                      category:
                        type: string
                      gallery:
                        type: array
                        items:
                            type: number
                      mostwanted:
                        type: boolean
                      stock:
                        type: number

        '400':
          description: Faltan campos necesarios para crear el producto, o el producto ya existe.
          content:
            application/json:
              schema:
                properties:
                  Mensaje: 
                    type: string
        
        '500':
          description: Server error
          content:
            application/json:
              schema:
                properties:
                  Mensaje: 
                    type: string
  /products/{id}:
    put:
      tags: 
        - productos
      summary: Edita un producto identificado con id
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          schema:
            type: string
          required: true
          description: Id del producto
      requestBody:
        required: true
        content:
          application/json:
            schema:
              properties:
                title:
                  type: string
                stock:
                  type: number
                description:
                  type: string
                fk_id_category:
                  type: string
                mostwanted:
                  type: number  

      responses: 
        '200':
          description: Producto editado con exito
          content:
            application/json:
              schema:
                properties:
                  list:
                    type: object
                    properties:
                      id:
                        type: number
                      title:
                        type: string
                      price:
                        type: number
                      description:
                        type: string
                      image:
                        type: string
                      category:
                        type: string
                      gallery:
                        type: array
                        items:
                          type: object 
                          properties:
                            picture_id:
                              type: number
                            picture_url:
                              type: string
                            description:
                              type: string
                      mostwanted:
                        type: boolean
                      stock:
                        type: number
                      
        '400':
          description: Bad request
          content:
            application/json:
              schema:
                properties:
                  msg:
                    type: string
                  ok:
                    type: boolean
                    
        '404':
          description: Producto a editar no encontrado
          content:
            application/json:
              schema:
                properties:
                  msg:
                    type: string
                  ok:
                    type: boolean
        '500':
          description: Server error
          content:
            application/json:
              schema:
                properties:
                  msg:
                    type: string
                  ok:
                    type: boolean

    delete:
      tags: 
        - productos
      summary: Elimina un producto identificado con id
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          schema:
            type: string
          required: true
          description: Id del producto
      
      responses: 
        '200':
          description: Producto eliminado con exito
          content:
            application/json:
              schema:
                properties:
                  list:
                    type: object
                    properties:
                      id:
                        type: number
                      title:
                        type: string
                      price:
                        type: number
                      description:
                        type: string
                      image:
                        type: string
                      category:
                        type: string
                      gallery:
                        type: array
                        items:
                          type: object 
                          properties:
                            picture_id:
                              type: number
                            picture_url:
                              type: number
                      mostwanted:
                        type: boolean
                      stock:
                        type: number
                      
        '400':
          description: Bad request
          content:
            application/json:
              schema:
                properties:
                  msg:
                    type: string
                  ok:
                    type: boolean
                    
        '404':
          description: Producto no encontrado
          content:
            application/json:
              schema:
                properties:
                  msg:
                    type: string
                  ok:
                    type: boolean
        '500':
          description: Server error
          content:
            application/json:
              schema:
                properties:
                  msg:
                    type: string
                  ok:
                    type: boolean
    get:
      tags: 
        - productos
      summary: Busca el producto identificado con id
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          schema:
            type: string
          required: true
          description: Id del producto
  
      
      responses: 
        '200':
          description: Producto encontrado con exito
          content:
            application/json:
              schema:
                properties:
                  list:
                    type: object
                    properties:
                      id:
                        type: number
                      title:
                        type: string
                      description:
                        type: string
                      category:
                        type: string
                      gallery:
                        type: array
                        items:
                          type: object 
                          properties:
                            picture_id:
                              type: number
                            picture_url:
                              type: number
                      mostwanted:
                        type: boolean
                      stock:
                        type: number
                  
        '404':
          description: Producto no encontrado
          content:
            application/json:
              schema:
                properties:
                  msg:
                    type: string                      
      
        '500':
          description: Server error
          content:
            application/json:
              schema:
                properties:
                  Mensaje: 
                    type: string
  /products/search:
    get:
      tags: 
        - productos
      summary: Busca los productos segun una palabra clave
      security:
        - bearerAuth: []
      parameters:
        - name: q
          in: query
          schema:
            type: string
          required: true
          description: Palabra clave a buscar
      
      responses: 
        '200':
          description: Productos encontrados con exito
          content:
            application/json:
              schema:
                properties:
                  list:
                    type: array
                    items:
                      type: object
                    properties:
                      id:
                        type: number
                      title:
                        type: string
                      description:
                        type: string
                      category:
                        type: string
                      gallery:
                        type: array
                        items:
                          type: object 
                          properties:
                            picture_id:
                              type: number
                            picture_url:
                              type: number
                      mostwanted:
                        type: boolean
                      stock:
                        type: number
                        
        '400':
          description: bad request
          content:
            application/json:
              schema:
                properties:
                  Mensaje: 
                    type: string
                              
        '404':
          description: Productos no encontrados
          content:
            application/json:
              schema:
                properties:
                  msg:
                    type: string

        '500':
          description: Server error
          content:
            application/json:
              schema:
                properties:
                  Mensaje: 
                    type: string
  /products/mostwanted:
    get:
      tags: 
        - productos
      summary: Lista los productos mas buscados
      security:
        - bearerAuth: []
      
      responses: 
        '200':
          description: Productos encontrados con exito
          content:
            application/json:
              schema:
                properties:
                  list:
                    type: object
                    properties:
                      id:
                        type: number
                      title:
                        type: string
                      description:
                        type: string
                      category:
                        type: string
                      gallery:
                        type: array
                        items:
                          type: object 
                          properties:
                            picture_id:
                              type: number
                            picture_url:
                              type: number
                      mostwanted:
                        type: boolean
                      stock:
                        type: number
                  
        '500':
          description:  server error
          content:
            application/json:
              schema:
                properties:
                  Mensaje: 
                    type: string
  /products/{id}/pictures:
    get:
      tags:
        - pictures
      summary: Lista las fotos del producto identificado con id
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          schema:
            type: number
          required: true
          description: Id del producto
          
      responses: 
        '200':
          description: Fotos listadas con exito
          content:
            application/json:
              schema:
                properties:
                  list:
                    type: array
                    items:
                      type: object
                    properties:
                      id:
                        type: number
                        
        '404':
          description: Producto no encontrado
          content:
            application/json:
              schema:
                properties:
                  msg:
                    type: string
                    
        '500':
          description: Server error
          content:
            application/json:
              schema:
                properties:
                  Mensaje: 
                    type: string
...